requirements.txt
streamlit
pymongo
google-generativeai
click==8.0.4